Issue #138 refactor: MonthCell관련 로직 MonthCellFactory.kt로 리펙토링
