+++ library/src/main/java/com/drunkenboys/ckscalendar/month/MonthPageAdapter.kt
@@ -8,12 +8,10 @@ import com.drunkenboys.ckscalendar.data.*
 import com.drunkenboys.ckscalendar.databinding.ItemMonthPageBinding
 import com.drunkenboys.ckscalendar.listener.OnDayClickListener
 import com.drunkenboys.ckscalendar.listener.OnDaySecondClickListener
-import com.drunkenboys.ckscalendar.utils.TimeUtils.parseDayWeekToDayType
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
-import java.time.DayOfWeek
 import java.time.LocalDate
 
 
@@ -32,6 +30,8 @@ class MonthPageAdapter : RecyclerView.Adapter<MonthPageAdapter.Holder>() {
 
     private val today = LocalDate.now()
     private var isFirstToday = true
+
+    private val monthCellFactory = MonthCellFactory()
 
     var onDayClickListener: OnDayClickListener? = null
     var onDaySecondClickListener: OnDaySecondClickListener? = null
@@ -66,9 +66,6 @@ class MonthPageAdapter : RecyclerView.Adapter<MonthPageAdapter.Holder>() {
 
     inner class Holder(private val binding: ItemMonthPageBinding) : RecyclerView.ViewHolder(binding.root) {
 
-        private val weekSize = 7
-        private val calendarFullSize = 42
-
         private val monthAdapter = MonthAdapter { pagePosition, selectPosition ->
             // TODO : 성능 문제로 개선 필요
             if (lastSelectPagePosition != pagePosition) {
@@ -93,39 +90,11 @@ class MonthPageAdapter : RecyclerView.Adapter<MonthPageAdapter.Holder>() {
         ) {
             CoroutineScope(Dispatchers.Default).launch {
                 val dates = mutableListOf<CalendarDate>()
-                val startMonth = item.startDate.monthValue
-                val startDay = item.startDate.dayOfWeek
-                val endMonth = item.endDate.monthValue
 
                 cachedCalendar[item.id]?.let {
                     dates.addAll(it)
                 } ?: run {
-                    (startMonth..endMonth).forEach { month ->
-                        when (month) {
-                            startMonth -> {
-                                // add Start Padding
-                                if (startDay != DayOfWeek.SUNDAY) {
-                                    dates.addAll(makePadding(startDay.ordinal))
-                                }
-
-                                // add Start Dates
-                                dates.addAll(makeDates(item.startDate, month))
-                            }
-                            else -> {
-                                // add Normal Dates
-                                dates.addAll(makeDates(item.endDate, month))
-                            }
-                        }
-                    }
-                    // add End Padding
-                    val weekPadding = 6 - dates.size % weekSize
-                    dates.addAll(makePadding(weekPadding))
-
-                    // add FullSize Padding
-                    if (dates.size < calendarFullSize) {
-                        val fullSizePadding = calendarFullSize - dates.size - 1
-                        dates.addAll(makePadding(fullSizePadding))
-                    }
+                    dates.addAll(monthCellFactory.makeCell(item))
                 }
 
                 // check and set recycle data
@@ -150,21 +119,6 @@ class MonthPageAdapter : RecyclerView.Adapter<MonthPageAdapter.Holder>() {
                     monthAdapter.onDayClickListener = onDayClick
                     monthAdapter.onDaySecondClickListener = onDaySecondClick
                 }
-            }
-        }
-
-        private fun makeDates(selectedDate: LocalDate, month: Int): List<CalendarDate> {
-            val monthDate = LocalDate.of(selectedDate.year, month, 1)
-            return (1..monthDate.lengthOfMonth()).map { day ->
-                val date = LocalDate.of(selectedDate.year, month, day)
-                val dayType = parseDayWeekToDayType(date.dayOfWeek)
-                CalendarDate(date, dayType)
-            }
-        }
-
-        private fun makePadding(paddingCount: Int): List<CalendarDate> {
-            return (0..paddingCount).map {
-                CalendarDate(LocalDate.now(), DayType.PADDING)
             }
         }
     }
